+++ src/spectrophotometer/index.ts
@@ -1,2 +1,4 @@
 export * from './ecros5300Profile';
 export * from './ecrosCli';
+export * from './ecrosProfiles';
+export * from './ecros5501Simulator';